Cleaning up

+++ mne/filter.py
@@ -439,11 +439,11 @@ def resample(x, up, down, npad=50, axis=0, window='flat'):
 
     """
     # make sure our arithmetic will work
-    ratio = np.float(up) / np.float(down)
+    ratio = float(up) / down
 
     # add some padding at beginning and end to make scipy's FFT
     # method work a little cleaner
-    pad_shape = np.array(x.shape, dtype='int64')
+    pad_shape = np.array(x.shape, dtype='int')
     pad_shape[axis] = npad
     x_len = x.shape[axis]
     keep = np.zeros((x_len), dtype='bool')
